Some Improvements
- UE related includes moved outside of THIRD_PARTY_INCLUDES

- Result Double variable converted to double

+++ Source/FF_DB_NANODBC/Public/NANODBC_Connection.h
@@ -30,7 +30,7 @@ public:
 	int64 Integer64 = 0;
 	
 	UPROPERTY(BlueprintReadOnly, meta = (ToolTip = ""))
-	float Double = (double)0.f;
+	double Double = (double)0.f;
 	
 	UPROPERTY(BlueprintReadOnly, meta = (ToolTip = ""))
 	bool Bool = false;
